pyproject.toml
[project]
name = "bvd"
version = "0.1.0"
description = "Breaking Version Detector - Extensible tool for detecting breaking dependency version changes"
readme = "README.md"
authors = [{ name = "Zack Kollar", email = "me@seedyrom.io" }]
requires-python = ">=3.10"
dependencies = ["packaging>=21.0", "python-hcl2>=4.0.0", "click>=8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
bvd = "bvd.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/bvd"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I"]

[tool.black]
line-length = 100

[dependency-groups]
dev = ["black>=25.1.0", "pytest>=8.4.1", "ruff>=0.12.5"]
